# doc/api.yaml
{"openapi":"3.0.0","paths":{"/album":{"get":{"operationId":"AlbumController_getAlbums","parameters":[],"responses":{"200":{"description":"Get all albums","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumEntity"}}}}}},"tags":["Album"]},"post":{"operationId":"AlbumController_createAlbum","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAlbumDTO"}}}},"responses":{"200":{"description":"Create new album","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumEntity"}}}}},"tags":["Album"]}},"/album/{id}":{"get":{"operationId":"AlbumController_getAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get single album by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumEntity"}}}}},"tags":["Album"]},"put":{"operationId":"AlbumController_updateAlbumInfo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAlbumDTO"}}}},"responses":{"200":{"description":"Update album info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumEntity"}}}}},"tags":["Album"]},"delete":{"operationId":"AlbumController_deleteAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete album","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumEntity"}}}},"204":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumEntity"}}}}},"tags":["Album"]}},"/artist":{"get":{"operationId":"ArtistController_getArtists","parameters":[],"responses":{"200":{"description":"Get all artists","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ArtistEntity"}}}}}},"tags":["Artist"]},"post":{"operationId":"ArtistController_createArtist","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArtistDTO"}}}},"responses":{"200":{"description":"Create new artist","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtistEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtistEntity"}}}}},"tags":["Artist"]}},"/artist/{id}":{"get":{"operationId":"ArtistController_getArtistById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get single artist by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtistEntity"}}}}},"tags":["Artist"]},"put":{"operationId":"ArtistController_updateArtistInfo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArtistDTO"}}}},"responses":{"200":{"description":"Update artist info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtistEntity"}}}}},"tags":["Artist"]},"delete":{"operationId":"ArtistController_deleteArtist","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete artist"},"204":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtistEntity"}}}}},"tags":["Artist"]}},"/track":{"get":{"operationId":"TrackController_getTracks","parameters":[],"responses":{"200":{"description":"Get all tracks","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrackEntity"}}}}}},"tags":["Track"]},"post":{"operationId":"TrackController_createTrack","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTrackDTO"}}}},"responses":{"200":{"description":"Create new track","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrackEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrackEntity"}}}}},"tags":["Track"]}},"/track/{id}":{"get":{"operationId":"TrackController_getTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get single track by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrackEntity"}}}}},"tags":["Track"]},"put":{"operationId":"TrackController_updateTrackInfo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTrackDTO"}}}},"responses":{"200":{"description":"Update track info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrackEntity"}}}}},"tags":["Track"]},"delete":{"operationId":"TrackController_deleteTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete track"},"204":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrackEntity"}}}}},"tags":["Track"]}},"/user":{"get":{"operationId":"UserController_getUsers","parameters":[],"responses":{"200":{"description":"Get all users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}},"tags":["User"]},"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}}},"responses":{"200":{"description":"Create user (following DTO should be used) CreateUserDto","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"201":{"description":""}},"tags":["User"]}},"/user/{id}":{"get":{"operationId":"UserController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get single user by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["User"]},"put":{"operationId":"UserController_updateUserPwd","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserPasswordDTO"}}}},"responses":{"200":{"description":"Update user's password UpdatePasswordDto (with attributes)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["User"]},"delete":{"operationId":"UserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete user"},"204":{"description":""}},"tags":["User"]}},"/favs":{"get":{"operationId":"FavoriteController_getFavorites","parameters":[],"responses":{"200":{"description":"get all favorites","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FavoritesEntity"}}}}}},"tags":["Favorites"]}},"/favs/{entityName}/{id}":{"post":{"operationId":"FavoriteController_addToFavs","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"entityName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Add entity to the favorites"},"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Favorites"]},"delete":{"operationId":"FavoriteController_removeFromFavs","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"entityName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete entity from the favorites"},"204":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Favorites"]}}},"info":{"title":"Assignment: REST Service","description":"The REST API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"AlbumEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"year":{"type":"number"},"artistId":{"type":"string","nullable":true}},"required":["id","name","year","artistId"]},"CreateAlbumDTO":{"type":"object","properties":{"name":{"type":"string"},"year":{"type":"number"},"artistId":{"type":"string"}},"required":["name","year","artistId"]},"ArtistEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"grammy":{"type":"boolean"}},"required":["id","name","grammy"]},"CreateArtistDTO":{"type":"object","properties":{"name":{"type":"string"},"grammy":{"type":"boolean"}},"required":["name","grammy"]},"UpdateArtistDTO":{"type":"object","properties":{}},"TrackEntity":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"artistId":{"type":"string","nullable":true},"albumId":{"type":"string","nullable":true},"duration":{"type":"number"}},"required":["id","name","artistId","albumId","duration"]},"CreateTrackDTO":{"type":"object","properties":{"name":{"type":"string"},"duration":{"type":"number"},"artistId":{"type":"string"},"albumId":{"type":"string"}},"required":["name","duration","artistId","albumId"]},"UpdateTrackDTO":{"type":"object","properties":{}},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"login":{"type":"string"},"password":{"type":"string"},"version":{"type":"number"},"createdAt":{"type":"number"},"updatedAt":{"type":"number"}},"required":["id","login","password","version","createdAt","updatedAt"]},"CreateUserDTO":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}},"required":["login","password"]},"UpdateUserPasswordDTO":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}},"required":["oldPassword","newPassword"]},"FavoritesEntity":{"type":"object","properties":{"artists":{"type":"array","items":{"type":"string"}},"albums":{"type":"array","items":{"type":"string"}},"tracks":{"type":"array","items":{"type":"string"}},"favoritesId":{"type":"string"}},"required":["artists","albums","tracks","favoritesId"]}}}}